perf: Optimize the way diff_per_col is handled
The way the diff_per_col results were handled
was inefficient. We used to cast the results
as JSON and then parse it again, instead of
directly using the Arrow Table objects.
The change required some work, because
Arrow Vectors have to be cast to Arrays
to be usable, and the BigInts have to be
cast to Numbers.